Update files based on review
