+++ app/src/main/java/io/intrepid/contest/screens/contestcreation/categorieslist/CategoriesListFragment.java
@@ -21,7 +21,7 @@ import io.intrepid.contest.screens.contestcreation.EditContestContract;
 import io.intrepid.contest.utils.dragdrop.SimpleItemTouchHelperCallback;
 
 
-public class CategoriesListFragment extends BaseFragment<CategoriesListPresenter> implements CategoriesListContract.View, ContestCreationFragment {
+public class CategoriesListFragment extends BaseFragment<CategoriesListPresenter> implements CategoriesContract.View, ContestCreationFragment {
     @BindView(R.id.generic_recycler_view)
     RecyclerView categoriesRecyclerView;
     private CategoryAdapter categoryAdapter;
