+++ app/src/main/java/io/intrepid/contest/screens/contestcreation/categorieslist/CategoryAdapter.java
@@ -20,10 +20,6 @@ public class CategoryAdapter extends RecyclerView.Adapter<CategoryViewHolder> im
     private Category exampleCategory;
     @LayoutRes private int rowItemLayout;
 
-    public CategoryAdapter(Context context) {
-        this(context, R.layout.category_card_row_item);
-    }
-
     public CategoryAdapter(Context context, @LayoutRes int rowLayout) {
         this.context = context;
         rowItemLayout = rowLayout;
