--- app/src/main/java/io/intrepid/contest/screens/contestjudging/submissionentrieslist/SubmissionEntriesContract.java
@@ -1,16 +1,0 @@
-package io.intrepid.contest.screens.contestjudging.submissionentrieslist;
-
-import java.util.List;
-
-import io.intrepid.contest.base.BaseContract;
-import io.intrepid.contest.models.Entry;
-
-class SubmissionEntriesContract {
-    public interface View extends BaseContract.View {
-        void showSubmissionList(List<Entry> submissionEntries);
-    }
-
-    public interface Presenter extends BaseContract.Presenter<View> {
-
-    }
-}
